Add s3 logic

+++ src/controllers/pratice-houses.js
@@ -3,6 +3,7 @@ import mysql from "mysql2/promise";
 import pool from "../db";
 import myRaw from "../utils/myRaw";
 import { constants } from "../utils/offers";
+import { deleteObjectByKey, deleteObjects } from "../utils/s3multer";
 
 const detailHouse = async (req, res) => {
   const { houseId } = req.params;
@@ -154,7 +155,6 @@ const createHouse = async (req, res) => {
     hourlyPrice,
     description,
     direction,
-    imageUrl,
     email,
     phoneNumber,
     rooms,
@@ -185,14 +185,13 @@ const createHouse = async (req, res) => {
     [myRaw.where.userId(req.user.recordId)]
   );
   const insertHouseSql = mysql.format(
-    `INSERT INTO practice_houses ( user_id, district_id, title, hourly_price, image_url, description, direction, email, phone_number ) 
-        VALUES ( ?,?,?,?,?,?,?,?,?)`,
+    `INSERT INTO practice_houses ( user_id, district_id, title, hourly_price,  description, direction, email, phone_number ) 
+        VALUES ( ?,?,?,?,?,?,?,?)`,
     [
       req.user.recordId,
       districtId,
       title,
       hourlyPrice,
-      imageUrl,
       description,
       direction,
       email,
@@ -209,28 +208,33 @@ const createHouse = async (req, res) => {
 
     await connection.query(insertHouseSql);
 
-    const [[{ lastInsertId }]] = await connection.query(
-      `SELECT LAST_INSERT_ID() AS lastInsertId`
+    const [[{ lastInsertHouseId }]] = await connection.query(
+      `SELECT LAST_INSERT_ID() AS lastInsertHouseId`
     );
 
+    const lastInsertRoomIds = [];
     for (const el of rooms) {
       const sql = mysql.format(
-        `INSERT INTO rooms (practice_house_id, name, instrument, hourly_price, image_url )
-            VALUES(?,?,?,?,?)`,
-        [lastInsertId, el.name, el.instrument, el.hourlyPrice, el.imageUrl]
+        `INSERT INTO rooms (practice_house_id, name, instrument, hourly_price )
+            VALUES(?,?,?,?)`,
+        [lastInsertHouseId, el.name, el.instrument, el.hourlyPrice]
       );
       await connection.query(sql);
+      const [[{ lastInsertRoomId }]] = await connection.query(
+        `SELECT LAST_INSERT_ID() AS lastInsertRoomId`
+      );
+      lastInsertRoomIds.push(lastInsertRoomId);
     }
     for (const el of facilityIds) {
       const sql = mysql.format(
         `INSERT INTO practice_houses_facility(practice_house_id, facility_id) 
             VALUES(?,?)`,
-        [lastInsertId, el]
+        [lastInsertHouseId, el]
       );
       await connection.query(sql);
     }
 
-    return res.status(201).json({ lastInsertId });
+    return res.status(201).json({ lastInsertHouseId, lastInsertRoomIds });
   } catch (error) {
     return res.status(403).json({ message: error.message });
   } finally {
@@ -243,7 +247,6 @@ const updateHouse = async (req, res) => {
   const {
     districtId,
     title,
-    imageUrl,
     description,
     direction,
     email,
@@ -277,12 +280,11 @@ const updateHouse = async (req, res) => {
   );
   const updateSql = mysql.format(
     `UPDATE practice_houses
-        SET district_id=?, title=?, image_url=?, description=?, email=?, phone_number=?, hourly_price=?, direction=?
+        SET district_id=?, title=?, description=?, email=?, phone_number=?, hourly_price=?, direction=?
         WHERE 1=1 ?`,
     [
       districtId,
       title,
-      imageUrl,
       description,
       email,
       phoneNumber,
@@ -317,39 +319,43 @@ const updateHouse = async (req, res) => {
 
     await connection.query(updateSql);
 
+    const Objects = [];
     for (const el of rooms) {
+      console.log(el);
       if (el.id) {
         if (el.isDeleted) {
+          if (!el.imageUrl) {
+            return res.status(403).json({ message: "OMISSION IS_DELETED" });
+          }
           const sql = mysql.format(
             `DELETE FROM rooms
                 WHERE 1=1 ?`,
             [myRaw.where.id(el.id)]
           );
           await connection.query(sql);
+
+          const Key = el.imageUrl.split(".com/")[1];
+          Objects.push({ Key });
         } else {
           const sql = mysql.format(
             `UPDATE rooms
-                SET name=?, instrument=?, hourly_price=?, image_url=?
+                SET name=?, instrument=?, hourly_price=?
                 WHERE 1=1 ?`,
-            [
-              el.name,
-              el.instrument,
-              el.hourlyPrice,
-              el.imageUrl,
-              myRaw.where.id(el.id),
-            ]
+            [el.name, el.instrument, el.hourlyPrice, myRaw.where.id(el.id)]
           );
           await connection.query(sql);
         }
       } else if (el.isNew) {
         const sql = mysql.format(
-          `INSERT INTO rooms(practice_house_id, name, instrument, hourly_price, image_url) 
-            VALUES(?,?,?,?,?)`,
-          [houseId, el.name, el.instrument, el.hourlyPrice, el.imageUrl]
+          `INSERT INTO rooms(practice_house_id, name, instrument, hourly_price) 
+            VALUES(?,?,?,?)`,
+          [houseId, el.name, el.instrument, el.hourlyPrice]
         );
         await connection.query(sql);
       }
     }
+
+    deleteObjects(Objects);
 
     for (const el of facilityIdsToCreate) {
       const sql = mysql.format(
@@ -380,9 +386,14 @@ const deleteHouse = async (req, res) => {
   const { houseId } = req.params;
 
   const checkExSql = mysql.format(
-    `SELECT id, user_id FROM practice_houses
+    `SELECT id, user_id, image_url FROM practice_houses
         WHERE 1=1 ?`,
     [myRaw.where.id(houseId)]
+  );
+  const selectRoomsSql = mysql.format(
+    `SELECT id, image_url FROM rooms
+        WHERE 1=1 ?`,
+    [myRaw.where.houseIdRefer(houseId)]
   );
   const deleteRoomsSql = mysql.format(
     `DELETE FROM rooms
@@ -408,10 +419,26 @@ const deleteHouse = async (req, res) => {
     } else if (exHouse.user_id !== req.user.recordId) {
       return res.status(403).json({ message: "INVALID USER" });
     }
+    const [exRooms] = await connection.query(selectRoomsSql);
 
     await connection.query(deleteRoomsSql);
     await connection.query(deletePracticeHousesFacilitiesSql);
     await connection.query(deleteHouseSql);
+
+    if (exHouse.image_url) {
+      const Key = exHouse.image_url.split(".com/")[1];
+      deleteObjectByKey(Key);
+    }
+
+    const Objects = [];
+    for (const el of exRooms) {
+      if (el.image_url) {
+        const Key = el.image_url.split(".com/")[1];
+        Objects.push({ Key });
+      }
+    }
+    deleteObjects(Objects);
+
     return res.status(204).end();
   } catch (error) {
     return res.status(403).json({ message: error.message });
@@ -460,6 +487,101 @@ const fulfillHouse = async (req, res) => {
   }
 };
 
+const putHouseImage = async (req, res) => {
+  const { houseId } = req.params;
+
+  if (!req.file) {
+    return res.status(403).json({ message: "OMISSION IN FILE" });
+  }
+  const { location } = req.file;
+
+  const checkExSql = mysql.format(
+    `SELECT id, user_id, image_url FROM practice_houses 
+        WHERE 1=1 ?`,
+    [myRaw.where.id(houseId)]
+  );
+  const updateSql = mysql.format(
+    `UPDATE practice_houses
+        SET image_url=?
+        WHERE 1=1 ?`,
+    [location, myRaw.where.id(houseId)]
+  );
+  const connection = await pool.getConnection(async (conn) => conn);
+
+  try {
+    const [[exHouse]] = await connection.query(checkExSql);
+    if (!exHouse) {
+      return res.status(403).json({ message: "RECORD NOT EXISTS" });
+    } else if (exHouse.user_id !== req.user.recordId) {
+      return res.status(403).json({ message: "INVALID USER" });
+    }
+
+    await connection.query(updateSql);
+
+    if (exHouse.image_url) {
+      const imageKey = exHouse.image_url.split(".com/")[1];
+      deleteObjectByKey(imageKey);
+    }
+
+    return res.status(200).json({ location });
+  } catch (error) {
+    return res.status(403).json({ message: error.message });
+  } finally {
+    connection.release();
+  }
+};
+
+const putRoomImages = async (req, res) => {
+  const { roomIds } = req.query;
+  if (!req.files) {
+    return res.status(403).json({ message: "OMISSION IN FILE" });
+  }
+  const locations = req.files.map((el) => el.location);
+  if (roomIds.length !== locations.length) {
+    return res.status(403).json({ message: "LENGTH DOES NOT MATCH" });
+  }
+
+  const checkExSql = mysql.format(
+    `SELECT id, image_url FROM rooms
+        WHERE 1=1 ?`,
+    [myRaw.where.ids(roomIds)]
+  );
+  const connection = await pool.getConnection(async (conn) => conn);
+
+  try {
+    const [exRooms] = await connection.query(checkExSql);
+    if (exRooms.length === 0) {
+      return res.status(403).json({ message: "RECORD NOT EXISTS" });
+    }
+
+    // !: needed to check exRooms more specifically
+    for (let i = 0; i < locations.length; i++) {
+      const updateSql = mysql.format(
+        `UPDATE rooms
+            SET image_url=?
+            WHERE 1=1 ?`,
+        [locations[i], myRaw.where.id(roomIds[i])]
+      );
+      await connection.query(updateSql);
+    }
+
+    const Objects = [];
+    for (const el of exRooms) {
+      if (el.image_url) {
+        const Key = el.image_url.split(".com/")[1];
+        Objects.push({ Key });
+      }
+    }
+    deleteObjects(Objects);
+
+    return res.status(200).json({ locations });
+  } catch (error) {
+    return res.status(403).json({ message: error.message });
+  } finally {
+    connection.release();
+  }
+};
+
 export {
   detailHouse,
   listHouses,
@@ -468,4 +590,6 @@ export {
   updateHouse,
   deleteHouse,
   fulfillHouse,
+  putHouseImage,
+  putRoomImages,
 };
